@@ -2,7 +2,7 @@
 Package simulation implements a full fledged Cosmos SDK application used for executing
 simulation test suites.
 
-Simulation App
+# Simulation App
 
 The SimApp type defines an application used for running extensive simulation
 testing suites. It contains all core modules, including governance, staking,
@@ -30,7 +30,7 @@ still be pseudo-randomly simulated.
 The simulation test suite also supports testing determinism and import/export
 functionality.
 
-Randomness
+# Randomness
 
 Currently, simulation uses a single seed (integer) as a source for a PRNG by
 which all random operations are executed from. Any call to the PRNG changes all
@@ -44,75 +44,74 @@ the simulation suite is expected to support a series of PRNGs that can be used
 uniquely per module and simulation component so that they will not effect each
 others state execution outcome.
 
-Usage
+# Usage
 
 To execute a completely pseudo-random simulation:
 
- $ go test -mod=readonly github.com/cosmos/cosmos-sdk/simapp \
-	-run=TestFullAppSimulation \
-	-Enabled=true \
-	-NumBlocks=100 \
-	-BlockSize=200 \
-	-Commit=true \
-	-Seed=99 \
-	-Period=5 \
-	-v -timeout 24h
+	 $ go test -mod=readonly github.com/cosmos/cosmos-sdk/simapp \
+		-run=TestFullAppSimulation \
+		-Enabled=true \
+		-NumBlocks=100 \
+		-BlockSize=200 \
+		-Commit=true \
+		-Seed=99 \
+		-Period=5 \
+		-v -timeout 24h
 
 To execute simulation from a genesis file:
 
- $ go test -mod=readonly github.com/cosmos/cosmos-sdk/simapp \
- 	-run=TestFullAppSimulation \
- 	-Enabled=true \
- 	-NumBlocks=100 \
- 	-BlockSize=200 \
- 	-Commit=true \
- 	-Seed=99 \
- 	-Period=5 \
-	-Genesis=/path/to/genesis.json \
- 	-v -timeout 24h
+	 $ go test -mod=readonly github.com/cosmos/cosmos-sdk/simapp \
+	 	-run=TestFullAppSimulation \
+	 	-Enabled=true \
+	 	-NumBlocks=100 \
+	 	-BlockSize=200 \
+	 	-Commit=true \
+	 	-Seed=99 \
+	 	-Period=5 \
+		-Genesis=/path/to/genesis.json \
+	 	-v -timeout 24h
 
 To execute simulation from a simulation params file:
 
- $ go test -mod=readonly github.com/cosmos/cosmos-sdk/simapp \
-	-run=TestFullAppSimulation \
-	-Enabled=true \
-	-NumBlocks=100 \
-	-BlockSize=200 \
-	-Commit=true \
-	-Seed=99 \
-	-Period=5 \
-	-Params=/path/to/params.json \
-	-v -timeout 24h
+	 $ go test -mod=readonly github.com/cosmos/cosmos-sdk/simapp \
+		-run=TestFullAppSimulation \
+		-Enabled=true \
+		-NumBlocks=100 \
+		-BlockSize=200 \
+		-Commit=true \
+		-Seed=99 \
+		-Period=5 \
+		-Params=/path/to/params.json \
+		-v -timeout 24h
 
 To export the simulation params to a file at a given block height:
 
- $ go test -mod=readonly github.com/cosmos/cosmos-sdk/simapp \
- 	-run=TestFullAppSimulation \
- 	-Enabled=true \
- 	-NumBlocks=100 \
- 	-BlockSize=200 \
- 	-Commit=true \
- 	-Seed=99 \
- 	-Period=5 \
-	-ExportParamsPath=/path/to/params.json \
-	-ExportParamsHeight=50 \
-	 -v -timeout 24h
-
+	 $ go test -mod=readonly github.com/cosmos/cosmos-sdk/simapp \
+	 	-run=TestFullAppSimulation \
+	 	-Enabled=true \
+	 	-NumBlocks=100 \
+	 	-BlockSize=200 \
+	 	-Commit=true \
+	 	-Seed=99 \
+	 	-Period=5 \
+		-ExportParamsPath=/path/to/params.json \
+		-ExportParamsHeight=50 \
+		 -v -timeout 24h
 
 To export the simulation app state (i.e genesis) to a file:
 
- $ go test -mod=readonly github.com/cosmos/cosmos-sdk/simapp \
- 	-run=TestFullAppSimulation \
- 	-Enabled=true \
- 	-NumBlocks=100 \
- 	-BlockSize=200 \
- 	-Commit=true \
- 	-Seed=99 \
- 	-Period=5 \
-	-ExportStatePath=/path/to/genesis.json \
-	 v -timeout 24h
+	 $ go test -mod=readonly github.com/cosmos/cosmos-sdk/simapp \
+	 	-run=TestFullAppSimulation \
+	 	-Enabled=true \
+	 	-NumBlocks=100 \
+	 	-BlockSize=200 \
+	 	-Commit=true \
+	 	-Seed=99 \
+	 	-Period=5 \
+		-ExportStatePath=/path/to/genesis.json \
+		 v -timeout 24h
 
-Params
+# Params
 
 Params that are provided to simulation from a JSON file are used to used to set
 both module parameters and simulation parameters. See sim_test.go for the full
