@@ -52,7 +52,8 @@ const AverageBlockTime = 6 * time.Second
 // * Config file for params
 // * whatever is needed for logging (tb + w rn)
 // OR: Could be a struct or something with options,
-//     to give caller ability to step through / instrument benchmarking if they wanted to, and add a cleanup function.
+//
+//	to give caller ability to step through / instrument benchmarking if they wanted to, and add a cleanup function.
 func SimulateFromSeed(
 	tb testing.TB,
 	w io.Writer,
