@@ -25,9 +25,9 @@ func getTestingMode(tb testing.TB) (testingMode bool, t *testing.T, b *testing.B
 // getBlockSize returns a block size as determined from the transition matrix.
 // It targets making average block size the provided parameter. The three
 // states it moves between are:
-//  - "over stuffed" blocks with average size of 2 * avgblocksize,
-//  - normal sized blocks, hitting avgBlocksize on average,
-//  - and empty blocks, with no txs / only txs scheduled from the past.
+//   - "over stuffed" blocks with average size of 2 * avgblocksize,
+//   - normal sized blocks, hitting avgBlocksize on average,
+//   - and empty blocks, with no txs / only txs scheduled from the past.
 func getBlockSize(simCtx *simtypes.SimCtx, params Params, lastBlockSizeState, avgBlockSize int) (state, blockSize int) {
 	r := simCtx.GetSeededRand("get block size")
 	// TODO: Make default blocksize transition matrix actually make the average
