@@ -3,9 +3,9 @@ The `mint` module is responsible for creating tokens in a
 flexible way to reward validators, incentivize providing pool
 liquidity, provide funds for Osmosis governance,
 and pay developers to maintain and improve Osmosis.
- - Denom minting; reduction and reserve ratio settings
- - Token distribution proportions
- - Epoch start distribution settings
+  - Denom minting; reduction and reserve ratio settings
+  - Token distribution proportions
+  - Epoch start distribution settings
 */
 package mint
 
