@@ -4,9 +4,9 @@ in the `incentives` module for every lock duration
 that exists in that pool. The `pool-incentives` module also takes
 the `pool_incentives` distributed from the `gov` module
 and distributes it to the various incentivized gauges.
- - Handles governance proposals impacting pool incentives.
- - Pool distribution and lockup infos queries.
- - Distributes incentives to LPs.
+  - Handles governance proposals impacting pool incentives.
+  - Pool distribution and lockup infos queries.
+  - Distributes incentives to LPs.
 */
 package pool_incentives
 
@@ -67,7 +67,7 @@ func (AppModuleBasic) ValidateGenesis(cdc codec.JSONCodec, config client.TxEncod
 	return types.ValidateGenesis(&data)
 }
 
-//---------------------------------------
+// ---------------------------------------
 // Interfaces.
 func (b AppModuleBasic) RegisterRESTRoutes(ctx client.Context, r *mux.Router) {
 	// noop
